refinement: rewriting tag_counts_array method to improve speed on tags list, tags rss feed, and basket homepages with tags list on.
Previous method was generating queries that would take close to 7 seconds to run through 15k tags, because the query was going through each zoom class, and checking every tag was in that basket. We also had some ugly limiting/sorting code after all tags were fetched (instead of doing that in the query).

Have hacked the taggable plugin to make use of a new basket_id column on the basket model. Now allows things like @current_basket.taggings. As a result, we can now sort, limit, and check basket permission all on the taggings table, and in the same SQL, without having to go to any item tables for information. Big speed benefits.