Add ability to CRUD comments on topics.
Previous to this comments were read-only i.e. we displayed any existing
comments in the system but users could not create new comments. This PR
re-enables that functionality.

The approach taken in this PR is to disturb the original code as little
as possible because it is very interconnected and to try to improve test
coverage as much as possible. It is hoped that in future with reasonable
test coverage then the (much needed) refactorings will be possible.

Summary of important changes to existing code:

* Remove references to #prepare_and_save_to_zoom as this method does not
  exist in the new Kete
* Remove reference to #renumber_all - this method came from
  `better_nested_set` gem (part of original Kete codebas) but is not
  required with the `awesome_nested_set` gem that replaces it.

Summary of changes to specs and seeds:

* Add more seeds to make the initial setup of the DB more complete. This
  work was required to create feature tests that would properly exercise
  creating comments on topics.
* Seed YAML files were simplified as much as possible as they were
  imported into `db/yaml/`.
* Some older seed YAML files that were present in both
  `db/old_bootstrap` and `db/yaml` were removed
* Some improvements were made factories as required. Mostly this was to
  make the factory output more "random" to make it easier to distinguish
  in test and error output.